UserList fixes
